backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
sentence-transformers
pypdf
groq
python-dotenv